requirements.txt
flask
flask-cors
google-cloud-aiplatform
google-cloud-firestore
gunicorn